+++ react/FilterNavigatorFlexible.js
@@ -4,10 +4,11 @@ import { useDevice } from 'vtex.device-detector'
 import { pathOr } from 'ramda'
 
 import FilterNavigator from './FilterNavigator'
+import FilterNavigatorContext from './components/FilterNavigatorContext'
 
 import styles from './searchResult.css'
 
-const withSearchPageContextProps = Component => ({ layout }) => {
+const withSearchPageContextProps = Component => ({ layout, initiallyCollapsed }) => {
   const {
     searchQuery,
     map,
@@ -16,14 +17,19 @@ const withSearchPageContextProps = Component => ({ layout }) => {
     hiddenFacets,
     filters,
     showFacets,
-    showContentLoader,
     preventRouteChange,
     facetsLoading,
   } = useSearchPage()
   const { isMobile } = useDevice()
 
   const facets = pathOr({}, ['data', 'facets'], searchQuery)
-  const { brands, priceRanges, specificationFilters, categoriesTrees } = facets
+  const {
+    brands,
+    priceRanges,
+    specificationFilters,
+    categoriesTrees,
+    queryArgs,
+  } = facets
 
   if (showFacets === false || !map) {
     return null
@@ -33,21 +39,24 @@ const withSearchPageContextProps = Component => ({ layout }) => {
     <div
       className={`${styles['filters--layout']} ${
         layout === 'desktop' && isMobile ? 'w-100 mh5' : ''
-      }`}
+        }`}
     >
-      <Component
-        preventRouteChange={preventRouteChange}
-        brands={brands}
-        params={params}
-        priceRange={priceRange}
-        priceRanges={priceRanges}
-        specificationFilters={specificationFilters}
-        tree={categoriesTrees}
-        loading={facetsLoading && showContentLoader}
-        filters={filters}
-        hiddenFacets={hiddenFacets}
-        layout={layout}
-      />
+      <FilterNavigatorContext.Provider value={queryArgs}>
+        <Component
+          preventRouteChange={preventRouteChange}
+          brands={brands}
+          params={params}
+          priceRange={priceRange}
+          priceRanges={priceRanges}
+          specificationFilters={specificationFilters}
+          tree={categoriesTrees}
+          loading={facetsLoading}
+          filters={filters}
+          hiddenFacets={hiddenFacets}
+          layout={layout}
+          initiallyCollapsed={initiallyCollapsed}
+        />
+      </FilterNavigatorContext.Provider>
     </div>
   )
 }