+++ react/components/CategoryFilter.js
@@ -1,11 +1,10 @@
 import classNames from 'classnames'
-import React, { useContext } from 'react'
+import React from 'react'
 import { injectIntl } from 'react-intl'
 import { IconClose } from 'vtex.styleguide'
-import { NoSSR } from 'vtex.render-runtime'
 import { useCssHandles } from 'vtex.css-handles'
+import { useFilterNavigator } from './FilterNavigatorContext'
 
-import QueryContext from './QueryContext'
 import Collapsible from './Collapsible'
 import CategoryItem from './CategoryItem'
 
@@ -38,8 +37,13 @@ const getSelectedCategories = rootCategory => {
   return selectedCategories
 }
 
-const CategoryFilter = ({ category, shallow = false, onCategorySelect }) => {
-  const { map } = useContext(QueryContext)
+const CategoryFilter = ({
+  category,
+  shallow = false,
+  onCategorySelect,
+  preventRouteChange,
+}) => {
+  const { map } = useFilterNavigator()
   const handles = useCssHandles(CSS_HANDLES)
 
   const selectedCategories = getSelectedCategories(category)
@@ -47,7 +51,7 @@ const CategoryFilter = ({ category, shallow = false, onCategorySelect }) => {
   const handleUnselectCategories = index => {
     const categoriesToRemove = selectedCategories.slice(index)
 
-    onCategorySelect(categoriesToRemove)
+    onCategorySelect(categoriesToRemove, preventRouteChange)
   }
 
   const lastSelectedCategory = selectedCategories[selectedCategories.length - 1]
@@ -60,7 +64,7 @@ const CategoryFilter = ({ category, shallow = false, onCategorySelect }) => {
     }
 
     if (shallow) {
-      onCategorySelect(category)
+      onCategorySelect()
     } else {
       // deselect root category
       handleUnselectCategories(0)
@@ -138,29 +142,28 @@ const CategoryFilter = ({ category, shallow = false, onCategorySelect }) => {
                 mt2: shallow,
               })}
             >
-              <NoSSR>
-                <Collapsible
-                  items={lastSelectedCategory.children}
-                  maxItems={8}
-                  threshold={2}
-                  linkClassName="ml3"
-                  openLabel="store/filter.more-categories"
-                  render={(childCategory, index) => (
-                    <CategoryItem
-                      key={childCategory.id}
-                      className={classNames({
-                        mt2: index === 0 && !shallow,
-                      })}
-                      onClick={() =>
-                        onCategorySelect(
-                          shallow ? [category, childCategory] : childCategory
-                        )
-                      }
-                      label={childCategory.name}
-                    />
-                  )}
-                />
-              </NoSSR>
+              <Collapsible
+                items={lastSelectedCategory.children}
+                maxItems={8}
+                threshold={2}
+                linkClassName="ml3"
+                openLabel="store/filter.more-categories"
+                render={(childCategory, index) => (
+                  <CategoryItem
+                    key={childCategory.id}
+                    className={classNames({
+                      mt2: index === 0 && !shallow,
+                    })}
+                    onClick={() =>
+                      onCategorySelect(
+                        shallow ? [category, childCategory] : childCategory,
+                        preventRouteChange
+                      )
+                    }
+                    label={childCategory.name}
+                  />
+                )}
+              />
             </div>
           )}
       </div>