+++ react/components/FacetItem.js
@@ -4,15 +4,18 @@ import { useCssHandles, applyModifiers } from 'vtex.css-handles'
 import classNames from 'classnames'
 
 import SettingsContext from './SettingsContext'
-import useFacetNavigation from '../hooks/useFacetNavigation'
 
 const CSS_HANDLES = ['filterItem']
 
-const FacetItem = ({ facet, className, preventRouteChange = false }) => {
+const FacetItem = ({
+  navigateToFacet,
+  facetTitle,
+  facet,
+  className,
+  preventRouteChange,
+}) => {
   const { showFacetQuantity } = useContext(SettingsContext)
   const handles = useCssHandles(CSS_HANDLES)
-  const navigateToFacet = useFacetNavigation()
-
   const classes = classNames(
     applyModifiers(handles.filterItem, facet.value),
     { [`${handles.filterItem}--selected`]: facet.selected },
@@ -32,7 +35,9 @@ const FacetItem = ({ facet, className, preventRouteChange = false }) => {
           showFacetQuantity ? `${facet.name} (${facet.quantity})` : facet.name
         }
         name={facet.name}
-        onChange={() => navigateToFacet(facet, preventRouteChange)}
+        onChange={() =>
+          navigateToFacet({ ...facet, title: facetTitle }, preventRouteChange)
+        }
         value={facet.name}
       />
     </div>