+++ react/constants.ts
@@ -1,0 +1,8 @@
+export const MAP_CATEGORY_CHAR = 'c'
+export const MAP_BRAND_CHAR = 'b'
+export const MAP_QUERY_KEY = 'map'
+export const MAP_VALUES_SEP = ','
+export const PATH_SEPARATOR = '/'
+export const SPACE_REPLACER = '-'
+export const FILTER_TITLE_SEP = '_'
+export const SPEC_FILTER = 'specificationFilter'